Python 3 updates

+++ mooclet_engine/engine/policies.py
@@ -74,7 +74,7 @@ def choose_policy_group(variables, context):
 			version_id = version.id
 		else:
 			version_id = version["id"]
-  
+
 		while (not allowed_versions.filter(pk=version_id).exists()) and count < context.get("maximum_allowed", 20):
 			count += 1
 			if Value.objects.filter(variable=grp_var, learner=context["learner"]).exists():
@@ -168,16 +168,16 @@ def choose_user_level_group(variables, context):
     Policy = apps.get_model('engine', 'Policy')
     PolicyParams = apps.get_model('engine', 'PolicyParameters')
     Mooclet = apps.get_model('engine', 'Mooclet')
-    
+
     var_name = str(context["mooclet"])+"_choose_user_level_group"
     grp_var, created = Variable.objects.get_or_create(name=var_name)
-    
+
     if "learner" not in context:
         return {"error": "please provide a learner ID"}
-    
+
     mooclet_values = Value.objects.filter(learner=context["learner"]).filter(~Q(variable=grp_var))
     mooclet_arms = Version.objects.filter(mooclet=context['mooclet'])
-    policy_parameters = context["policy_parameters"].parameters         
+    policy_parameters = context["policy_parameters"].parameters
     policies = list(policy_parameters["policies"])
     reward_name = policy_parameters["reward"]
     context_names = list(policy_parameters["contexts"])
@@ -209,7 +209,7 @@ def choose_user_level_group(variables, context):
         for index, policy_param in enumerate(policies):
             policy = Policy.objects.get(name=policy_param.get('name'))
             parameters = policy_param.get('parameters')
-            
+
             if policy.name in ['ts_configurable', 'ts_configurable_with_r']:
                 posteriors_data = parameters.get('current_posteriors', None)
                 if posteriors_data:
@@ -224,25 +224,25 @@ def choose_user_level_group(variables, context):
 
             PolicyParams.objects.create(
                 mooclet=user_mooclet,
-                policy=policy, 
+                policy=policy,
                 parameters=policy_param.get('parameters'),
                 latest_update=datetime.datetime.now()
             )
-            
+
             print("policy created: {}".format(policy.name))
-            
+
             # link the first policy to new mooclet
             if index == 0:
                 user_mooclet.policy = policy
                 user_mooclet.save()
-        
+
         # Create history record to link current mooclet and sub-mooclet
         Value.objects.create(variable=grp_var, learner=context["learner"], text=new_mooclet_name)
-    
+
     # update mooclet id to all current values
     mooclet_values.update(mooclet=user_mooclet)
     print("policy value updated: {}".format(mooclet_values))
-    
+
     # Note: assume sub-mooclets share the same arm name as the current mooclet
     for arm in mooclet_arms:
         version_values = mooclet_values.filter(variable__name='version', version=arm)
@@ -254,9 +254,9 @@ def choose_user_level_group(variables, context):
 
         # update arm id to all reward values
         reward_values.update(version=user_arm)
-        
+
         print("arm updated: {}".format(arm.pk))
-    
+
     version = user_mooclet.run(context={"learner":context["learner"]})
     if type(version) != dict:
         version_dict = model_to_dict(version)
@@ -634,11 +634,11 @@ def if_then_rules(variables, context):
 	version_to_show = choice(context['mooclet'].version_set.all())
 	if_then_rules_ur, created = Variable.objects.get_or_create(name="IF_THEN_RULES_UR")
 	Value.objects.create(
-		variable=if_then_rules_ur, 
-		value=1.0, 
-		text="IF_THEN_RULES_UR", 
-		learner=context["learner"], 
-		mooclet=context["mooclet"], 
+		variable=if_then_rules_ur,
+		value=1.0,
+		text="IF_THEN_RULES_UR",
+		learner=context["learner"],
+		mooclet=context["mooclet"],
 		version=version_to_show
 	)
 	return version_to_show
@@ -1470,7 +1470,7 @@ def ts_configurable(variables, context):
 				sum_rewards = 0
 
 
-			success_update = (sum_rewards - rating_count * min_rating) / (max_rating - min_rating)
+			success_update = float(sum_rewards - rating_count * min_rating) / float(max_rating - min_rating)
 			successes = success_update
 			failures = rating_count - success_update
 
@@ -1617,11 +1617,11 @@ def ts_configurable_with_r(variables, context):
 		uniform_threshold = 0
 
 	current_enrolled = Value.objects.filter(
-		variable__name="version", 
-		mooclet=context["mooclet"], 
+		variable__name="version",
+		mooclet=context["mooclet"],
 		policy__name="ts_configurable_with_r"
 	)
-	
+
 	if uniform_threshold > 0 and current_enrolled.count() <= policy_parameters["uniform_threshold"]:
 		ur_or_ts, created = Variable.objects.get_or_create(name="UR_or_TS")
 		version_to_show = choice(context['mooclet'].version_set.all())
@@ -1654,7 +1654,7 @@ def ts_configurable_with_r(variables, context):
 			else:
 				rating_count = 0
 				sum_rewards = 0
-    
+
 			update_para = R.r(
 				r'''
 				function(para){
@@ -1669,7 +1669,7 @@ def ts_configurable_with_r(variables, context):
 				}
 				'''
 			)
-   
+
 			as_numeric = R.r['as.numeric']
 			para_vec = R.FloatVector((0, 0, rating_count, sum_rewards))
 			para = update_para(para_vec)
@@ -1691,7 +1691,7 @@ def ts_configurable_with_r(variables, context):
 	version_dict = {}
 	for version in current_posteriors:
 		version_dict[Version.objects.get(id=version)] = {
-      		"successes": current_posteriors[version]["successes"] + prior_success, 
+      		"successes": current_posteriors[version]["successes"] + prior_success,
         	"failures":  current_posteriors[version]["failures"] + prior_failure
         }
 
@@ -1735,13 +1735,13 @@ def toptwo_ts_sample_with_r(toptwots_thresh, versions_dict, context):
 			algorithm_para <- para[5]
 			draw_a <- rbeta(1, sa, fa)
 			draw_b <- rbeta(1, sb, fb)
-			
+
 			is_ur <- 0
 			if (runif(1) < algorithm_para){
 				draw_a <- runif(1)
 				draw_b <- runif(1)
 				is_ur <- 1
-			} 
+			}
 			arm <- 1
 			if(draw_a < draw_b){
 				arm <- 2
@@ -1768,7 +1768,7 @@ def toptwo_ts_sample_with_r(toptwots_thresh, versions_dict, context):
 	ur_or_ts, created = Variable.objects.get_or_create(name="UR_or_TS")
 	if is_ur == 1:
 		Value.objects.create(
-      		variable=ur_or_ts, value=0.0, text="UR", 
+      		variable=ur_or_ts, value=0.0, text="UR",
             learner=context["learner"], mooclet=context["mooclet"],
 			version=version_to_show
    		)
@@ -1778,7 +1778,7 @@ def toptwo_ts_sample_with_r(toptwots_thresh, versions_dict, context):
 	else:
 		#log policy chosen
 		Value.objects.create(
-      		variable=ur_or_ts, value=1.0, text="TS", 
+      		variable=ur_or_ts, value=1.0, text="TS",
             learner=context["learner"], mooclet=context["mooclet"],
 			version=version_to_show
 		)
@@ -1810,13 +1810,13 @@ def ts_probclip_sample_with_r(tsprobclip_thresh, versions_dict, context):
 			algorithm_para <- para[5]
 			draw_a <- rbeta(1, sa, fa)
 			draw_b <- rbeta(1, sb, fb)
-		
+
 			a <- rbeta(1000, sa, fa)
 			b <- rbeta(1000, sb, fb)
 			pab <- mean(a > b)
-			
+
 			draw_a <- median(c(algorithm_para, 1 - algorithm_para, pab))
-			
+
 			arm <- 1
 			if(draw_a < draw_b) {
 				arm <- 2
@@ -1907,7 +1907,7 @@ def ts_postdiff_sample_with_r(tspostdiff_thresh, versions_dict, context):
 	ur_or_ts, created = Variable.objects.get_or_create(name="UR_or_TS")
 	if is_ur == 1:
 		Value.objects.create(
-      		variable=ur_or_ts, value=0.0, text="UR", 
+      		variable=ur_or_ts, value=0.0, text="UR",
             learner=context["learner"], mooclet=context["mooclet"],
 			version=version_to_show
    		)
@@ -1917,7 +1917,7 @@ def ts_postdiff_sample_with_r(tspostdiff_thresh, versions_dict, context):
 	else:
 		#log policy chosen
 		Value.objects.create(
-      		variable=ur_or_ts, value=1.0, text="TS", 
+      		variable=ur_or_ts, value=1.0, text="TS",
             learner=context["learner"], mooclet=context["mooclet"],
 			version=version_to_show
 		)
@@ -1948,7 +1948,7 @@ def ts_sample_with_r(versions_dict, context):
 			fb <- para[4]
 			draw_a <- rbeta(1, sa, fa)
 			draw_b <- rbeta(1, sb, fb)
-			
+
 			arm <- 1
 			if(draw_a < draw_b) {
 				arm <- 2
